new css, moved elemewnts around and more alignment
