+++ src/App.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import './App.css';
-import Logo from './components/logo'
 import Mailinglist from './components/mailinglist'
 import Graphic from './components/graphic'
 import Navbar from "./components/nav";
@@ -10,8 +9,7 @@ function App() {
   return (
     <div className="App">
       <header className="App-body">
-      <Logo></Logo>
-      <Navbar></Navbar>
+      <Navbar> </Navbar>
         <Graphic></Graphic>
         <Mailinglist></Mailinglist>
       </header>
