+++ src/components/graphic.jsx
@@ -1,10 +1,10 @@
 import React from "react";
-import benchpeople from "../assets/images/benchpeople.jpg"
+import benchpeople from "../assets/images/undraw_couplekiki.svg"
 
-export default function Graphic( props) {
+export default function Graphic(props) {
     return (
         <div className="graphiccontainer">
         <img className="graphic" src={benchpeople} alt="people on a bench"></img>
         </div>
     )
-}
+    }
