+++ src/components/mailinglist.jsx
@@ -4,8 +4,9 @@ export default function Mailing(props) {
     return (
         <div className="mailinglist">
             <p className="slogan">Get a fresh take on what you didn't learn in sex-ed.</p>
-            <a  href="form.html"><button className="mailinglistbtn">Join our mailing list</button></a>
-            <a className="quizlink" href="https://kristawhite122148.typeform.com/to/hNxxl3 " target="_blank" rel="noopener noreferrer">or take a quick Survey</a>
+            <a href="./form.html"><button className="mailinglistbtn">Join our mailing list</button></a>
+            <br></br>
+            <a className="quizlink" href="https://kristawhite122148.typeform.com/to/hNxxl3 " target="_blank" rel="noopener noreferrer">...Or take a quick survey</a>
             
         </div>
     )
